// src/types/generated/models.ts
export type Access = {
    data: AccessData
    delete: boolean
    externalize: boolean
    manage: boolean
    read: boolean
    update: boolean
    write: boolean
}
export type AccessData = {
    read: boolean
    write: boolean
}
// ...
    export enum type {
        PERSONAL_ACCESS_TOKEN_V1 = 'PERSONAL_ACCESS_TOKEN_V1',
        PERSONAL_ACCESS_TOKEN_V2 = 'PERSONAL_ACCESS_TOKEN_V2',
    }
// ...
export type Attribute = {
    access: Access
    attributeValues: Array<AttributeValue>
    categoryAttribute: boolean
    categoryOptionAttribute: boolean
    categoryOptionComboAttribute: boolean
    categoryOptionGroupAttribute: boolean
    categoryOptionGroupSetAttribute: boolean
    code: string
    constantAttribute: boolean
    created: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    createdBy?: {
        id: string
    }
    dataElementAttribute: boolean
    dataElementGroupAttribute: boolean
    dataElementGroupSetAttribute: boolean
    dataSetAttribute: boolean
    description: string
    displayDescription: string
    displayFormName: string
    displayName: string
    displayShortName: string
    documentAttribute: boolean
    eventChartAttribute: boolean
    eventReportAttribute: boolean
    favorite: boolean
    favorites: Array<string>
    formName: string
    href: string
    id: string
    indicatorAttribute: boolean
    indicatorGroupAttribute: boolean
    lastUpdated: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    lastUpdatedBy?: {
        id: string
    }
    legendSetAttribute: boolean
    mandatory: boolean
    mapAttribute: boolean
    name: string
    objectTypes: Array<string>
    optionAttribute: boolean
    /**
     * A UID reference to a OptionSet
     * (Java name `org.hisp.dhis.option.OptionSet`)
     */
    optionSet?: {
        id: string
    }
    optionSetAttribute: boolean
    organisationUnitAttribute: boolean
    organisationUnitGroupAttribute: boolean
    organisationUnitGroupSetAttribute: boolean
    programAttribute: boolean
    programIndicatorAttribute: boolean
    programStageAttribute: boolean
    relationshipTypeAttribute: boolean
    sectionAttribute: boolean
    sharing: Sharing
    shortName: string
    sortOrder: number
    sqlViewAttribute: boolean
    trackedEntityAttributeAttribute: boolean
    trackedEntityTypeAttribute: boolean
    translations: Array<Translation>
    unique: boolean
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    user?: {
        id: string
    }
    userAttribute: boolean
    userGroupAttribute: boolean
    validationRuleAttribute: boolean
    validationRuleGroupAttribute: boolean
    valueType: Attribute.valueType
    visualizationAttribute: boolean
}

export namespace Attribute {
    export enum valueType {
        LIST = 'LIST',
        TEXT = 'TEXT',
        LONG_TEXT = 'LONG_TEXT',
        MULTI_TEXT = 'MULTI_TEXT',
        LETTER = 'LETTER',
        PHONE_NUMBER = 'PHONE_NUMBER',
        EMAIL = 'EMAIL',
        BOOLEAN = 'BOOLEAN',
        TRUE_ONLY = 'TRUE_ONLY',
        DATE = 'DATE',
        DATETIME = 'DATETIME',
        TIME = 'TIME',
        NUMBER = 'NUMBER',
        UNIT_INTERVAL = 'UNIT_INTERVAL',
        PERCENTAGE = 'PERCENTAGE',
        INTEGER = 'INTEGER',
        INTEGER_POSITIVE = 'INTEGER_POSITIVE',
        INTEGER_NEGATIVE = 'INTEGER_NEGATIVE',
        INTEGER_ZERO_OR_POSITIVE = 'INTEGER_ZERO_OR_POSITIVE',
        TRACKER_ASSOCIATE = 'TRACKER_ASSOCIATE',
        USERNAME = 'USERNAME',
        COORDINATE = 'COORDINATE',
        ORGANISATION_UNIT = 'ORGANISATION_UNIT',
        REFERENCE = 'REFERENCE',
        AGE = 'AGE',
        URL = 'URL',
        FILE_RESOURCE = 'FILE_RESOURCE',
        IMAGE = 'IMAGE',
        GEOJSON = 'GEOJSON',
    }
}
// ...
export type AttributeValue = {
    /**
     * A UID reference to a Attribute
     * (Java name `org.hisp.dhis.attribute.Attribute`)
     */
    attribute?: {
        id: string
    }
    value: string
}
// ...
export type Sharing = {
    external: boolean
    owner: string
    public: string
    userGroups: Record<string, UserGroupAccess>
    users: Record<string, UserAccess>
}
// ...
export type Translation = {
    locale: string
    property: string
    value: string
}
// ...
export type UserAccess = {
    access: string
    displayName: string
    id: string
}
// ...
export type UserGroupAccess = {
    access: string
    displayName: string
    id: string
}